+++ llvm/include/llvm/Support/TypeSize.h
@@ -322,8 +322,8 @@ public:
   static constexpr TypeSize getFixed(ScalarTy ExactSize) {
     return TypeSize(ExactSize, false);
   }
-  static constexpr TypeSize getScalable(ScalarTy MinimunSize) {
-    return TypeSize(MinimunSize, true);
+  static constexpr TypeSize getScalable(ScalarTy MinimumSize) {
+    return TypeSize(MinimumSize, true);
   }
   static constexpr TypeSize get(ScalarTy Quantity, bool Scalable) {
     return TypeSize(Quantity, Scalable);